feat: MusicPanel compact 레이아웃 추가
